src/serving/requirements-docker.txt
fastapi
uvicorn[standard]
pydantic
redis
python-dotenv

numpy
pandas
scikit-learn
lightgbm
shap

mlflow
dagshub
prometheus-client==0.20.0
